% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.lmerModLmerTest}
\alias{augment.lmerModLmerTest}
\title{Augment data with information from an object}
\usage{
augment.lmerModLmerTest(x, data, ...)
}
\arguments{
\item{x}{Model object of class lmerModLmerTest with information to append
to observations.}

\item{data}{observations to be augmented}

\item{...}{Addition arguments to augment method.}
}
\value{
A \link[tibble]{tibble} with information about data points.
}
\description{
Augment data with information from an object
}
