% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment_with_p.flag}
\alias{augment_with_p.flag}
\title{Add significance flag}
\usage{
augment_with_p.flag(data, p.value = "p.value")
}
\arguments{
\item{data}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} containing a column of p-values}

\item{p.value}{name of the p-value column}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with an additional character column named p.flag
}
\description{
Add significance flag
}
\examples{
model <- lm(mpg ~ vs, data = mtcars)
model_summary <- broom::tidy(model)
augment_with_p.flag(model_summary)
}
