% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_complete.cases}
\alias{filter_complete.cases}
\title{Filter Complete Cases}
\usage{
filter_complete.cases(data)
}
\arguments{
\item{data}{dataset to filter for complete cases}
}
\value{
data with only complete cases
}
\description{
Filter Complete Cases
}
\examples{
tribble(
  ~x, ~y,
  11, 12,
  21, NA
) \%>\%
filter_complete.cases()
}
