% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{formula_terms}
\alias{formula_terms}
\title{Formula Terms}
\usage{
formula_terms(data, formula)
}
\arguments{
\item{data}{dataset corresponding to the formula}

\item{formula}{the model formula as character}
}
\value{
character vector of terms in the formula
}
\description{
Formula Terms
}
\examples{
formula_terms(mtcars, "mpg ~ cyl")
}
