% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{keep_conf.int}
\alias{keep_conf.int}
\title{Keep Confidence Interval Columns}
\usage{
keep_conf.int(data, conf.int)
}
\arguments{
\item{data}{broom tidy dataframe}

\item{conf.int}{logical indicating whether to keep \code{conf.low} and \code{conf.high}}
}
\value{
the broom tidy dataframe with or without the confidence interval
columns
}
\description{
Keep Confidence Interval Columns
}
\examples{
\dontrun{
lm(mpg ~ vs, data = mtcars) \%>\%
  tidy(conf.int = TRUE) \%>\%
  keep_conf.int(FALSE)
}
}
