% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_cor}
\alias{lp_cor}
\title{Multiple comparison correlation tests}
\usage{
lp_cor(
  data,
  value,
  name,
  id,
  conf.int = TRUE,
  conf.level = 0.95,
  descriptives = TRUE,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response, group, and unit variables}

\item{value}{response variable corresponding to name}

\item{name}{source of the value}

\item{id}{sampling unit variable (e.g., persons)}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{descriptives}{= logical to add descriptive statistics or not}

\item{...}{additional parameters passed onto \code{\link[stats:cor.test]{stats::cor.test()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with information about the model components;
one model per row
}
\description{
Test for association between paired samples, using one of
Pearson's product moment correlation coefficient, Kendall's tau or
Spearman's rho for each pair of variables.
}
\note{
Tests are performed on pairwise complete data.
}
\examples{
lp_cor(sleep, extra, group, ID)
}
