% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_glmer}
\alias{lp_glmer}
\title{Generalized linear mixed-effects models fit to all pairs of group predictor}
\usage{
lp_glmer(
  data,
  formula,
  response,
  class,
  family = stats::gaussian(),
  descriptives = TRUE,
  exponentiate = FALSE,
  conf.int = FALSE,
  conf.level = 0.95,
  conf.method = "Wald",
  profile = NULL,
  ...
)
}
\arguments{
\item{data}{data frame containing the variables named in formula}

\item{formula}{a two-sided linear formula object describing both the
fixed-effects and random-effects part of the model, with the response on the
left of a ~ operator and the terms, separated by + operators, on the right.
Random-effects terms are distinguished by vertical bars ("|") separating
expressions for design matrices from grouping factors.}

\item{response}{response variable}

\item{class}{classification variable}

\item{family}{a GLM family, see \code{\link[stats:glm]{glm}} and
\code{\link[stats:family]{family}}.}

\item{descriptives}{logical to add raw descriptive statistics or not}

\item{exponentiate}{whether to exponentiate the fixed-effect coefficient
estimates and confidence intervals (common for logistic regression); if
\code{TRUE}, also scales the standard errors by the exponentiated
coefficient, transforming them to the new scale}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{conf.method}{method for computing confidence intervals
(see \code{lme4::confint.merMod})}

\item{profile}{pre-computed profile object, for speed when using
\code{conf.method="profile"}}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with one row for each
}
\description{
Generalized linear mixed-effects models fit to all pairs of group predictor
}
\examples{
lp_glmer(
  lme4::cake,
  "angle ~ recipe * temperature + (1 | recipe:replicate)",
  angle,
  recipe,
  poisson
)
}
