% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_riskratio}
\alias{lp_riskratio}
\title{Multiple comparison risk ratios}
\usage{
lp_riskratio(
  data,
  response,
  class,
  conf.int = TRUE,
  conf.level = 0.95,
  descriptives = TRUE,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response and exposure variables}

\item{response}{response variable}

\item{class}{exposure variable}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{descriptives}{= logical to add descriptive statistics or not}

\item{...}{additional parameters ultimately passed onto \code{\link[stats:glm]{stats::glm()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with computed odds ratios and associated
information
}
\description{
Computes risk ratios for each pair of exposures from tidy data.
}
\examples{
tribble(
  ~admit, ~rank, ~n,
  0, 1, 28,
  0, 2, 97,
  0, 3, 93,
  0, 4, 55,
  1, 1, 33,
  1, 2, 54,
  1, 3, 28,
  1, 4, 12
) \%>\%
  mutate(rank = as.character(rank)) \%>\%
  uncount(n) \%>\%
  lp_riskratio(admit, rank)
}
