% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_ttest_dep}
\alias{lp_ttest_dep}
\title{Multiple comparison dependent sample t-tests}
\usage{
lp_ttest_dep(
  data,
  response,
  class,
  unit,
  conf.int = TRUE,
  conf.level = 0.95,
  descriptives = TRUE,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response, group, and unit variables}

\item{response}{response variable}

\item{class}{class variable}

\item{unit}{sampling unit variable (e.g., persons)}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{descriptives}{= logical to add descriptive statistics or not}

\item{...}{additional parameters passed onto \code{\link[stats:t.test]{stats::t.test()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with information about the model components;
one model per row
}
\description{
Performs paired t-tests for each pair of groups where
observations are paired by unit from tidy data.
}
\note{
Tests are performed on pairwise complete data.
}
\examples{
lp_ttest_dep(sleep, extra, group, ID)
}
