% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_ttest_one}
\alias{lp_ttest_one}
\title{Multiple comparison one-sample t-tests}
\usage{
lp_ttest_one(
  data,
  response,
  class,
  conf.int = TRUE,
  conf.level = 0.95,
  descriptives = TRUE,
  mu = 0,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response and group variables}

\item{response}{response variable}

\item{class}{class variable}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{descriptives}{= logical to add descriptive statistics or not}

\item{mu}{a number indicating the true value of the mean}

\item{...}{additional parameters passed onto \code{\link[stats:t.test]{stats::t.test()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with information about the model components; one
model per row
}
\description{
Performs one-sample t-tests for each group from tidy data.
}
\examples{
lp_ttest_one(mtcars, mpg, cyl, 10)
}
