% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_ttest_two}
\alias{lp_ttest_two}
\title{Multiple comparison independent sample t-tests}
\usage{
lp_ttest_two(
  data,
  response,
  class,
  conf.int = TRUE,
  conf.level = 0.95,
  descriptives = TRUE,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response and group variables}

\item{response}{response variable}

\item{class}{classification variable}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{descriptives}{logical to add descriptive statistics or not}

\item{...}{additional parameters passed onto \code{\link[stats:t.test]{stats::t.test()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with information about the model components;
one model per row
}
\description{
Performs independent sample t-tests for each pair of groups
from tidy data.
}
\examples{
lp_ttest_two(mtcars, mpg, cyl)
}
