% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_via_cor_sig}
\alias{select_via_cor_sig}
\title{Subset columns using a correlation p-value}
\usage{
select_via_cor_sig(data, referent, p.value, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{referent}{name of the common variable with which to compute
correlations}

\item{p.value}{the maximum desired p-value}

\item{...}{additional arguments passed to \code{\link[stats:cor.test]{stats::cor.test()}}}
}
\value{
\code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Select variables in a data frame based on the p-value of the
correlation with one of the variables.
}
\examples{
select_via_cor_sig(mtcars, mpg, 0.001)
}
