% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_with_sided_ci}
\alias{tidy_with_sided_ci}
\title{One-sided Confidence Intervals}
\usage{
tidy_with_sided_ci(
  x,
  conf.level = 0.95,
  exponentiate = FALSE,
  alternatives = "two.sided",
  ...
)
}
\arguments{
\item{x}{A model object}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults
to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to
FALSE.}

\item{alternatives}{a character vector specifying the alternative
hypothesis for the test of each coefficient. Elements must be one of
"two.sided", "greater" or "less". Elements will be recycled to match the
number of tests.}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{\link[broom:reexports]{broom::tidy()}} object with sided confidence intervals
}
\description{
One-sided Confidence Intervals
}
\examples{
## alternative can be specified for each estimate
model <- lm(mpg ~ vs, data = mtcars)
tidy_with_sided_ci(model, alternatives = c("two.sided", "greater"))

## not all `broom::tidy()` implement the `conf.level` argument
bind_rows(
  # `tidy.htest()`
  broom::tidy(
    t.test(
      mpg ~ vs,
      data = mutate(mtcars, vs = -(vs - 1)),
      alternative = "greater",
      var.equal = TRUE,
      conf.level = 0.90 # ignored in `tidy.htest()`
    )
  ) \%>\%
    select(
      estimate,
      statistic,
      p.value,
      conf.low,
      conf.high,
      alternative
    ),

  # `tidy.lm()`
  tidy_with_sided_ci(
    lm(mpg ~ vs, data = mtcars),
    conf.level = 0.90, # not ignored in `tidy.lm()`
    alternatives = "greater"
  ) \%>\%
    filter(term == "vs") \%>\%
    select(
      estimate,
      statistic,
      p.value,
      conf.low,
      conf.high,
      alternative
    )
)
}
